# Conflicts:
#	src/main/kotlin/com/github/noamm9/features/impl/general/storageoverlay/StorageOverlayScreen.kt
#	src/main/kotlin/com/github/noamm9/utils/items/EtherwarpHelper.kt

+++ src/main/kotlin/com/github/noamm9/utils/items/EtherwarpHelper.kt
@@ -44,7 +44,7 @@ object EtherwarpHelper {
         if (stack.skyblockId.equalsOneOf("ASPECT_OF_THE_VOID", "ASPECT_OF_THE_END")) {
             val nbt = stack.customData
             if (nbt.getByte("ethermerge").orElse(0) != 1.toByte()) return null
-            val tuners = nbt.getByte("tuned_transmission").getOrDefault(0)
+            val tuners = nbt.getByte("tuned_transmission").getOrDefault(0).toInt()
             return 57.0 + tuners
         }
         return null
